some more work removed id.go
